backend/requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
tortoise-orm==0.20.0
aiomysql==0.2.0
asyncmy==0.2.11

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# 验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 文档转换
python-docx==1.1.0
pdfplumber==0.10.3
python-pptx==0.6.23
html2text==2020.1.16
markdown==3.5.1

# 工具
python-dotenv==1.0.0
aiofiles==23.2.1
